abstract the logic out, export, use for lrep
