+++ packages/axios-extension/src/abap/message.ts
@@ -1,5 +1,6 @@
 import type { Logger } from '@sap-ux/logger';
 import { URL } from 'url';
+import { isAxiosError } from 'axios';
 
 /**
  * Message detail object
@@ -178,3 +179,55 @@ export const prettyPrintTimeInMs = (ms: number): string => {
         return `${ms / 1000} seconds`;
     }
 };
+
+/**
+ * Log errors more user friendly if it is a standard Gateway error.
+ *
+ * @param e error thrown by Axios after sending a request
+ * @param e.error error from Axios
+ * @param e.host hostname
+ * @param e.log logger to be used
+ * @param e.isDest optional destination flag
+ */
+export function logError({
+    error,
+    host,
+    log,
+    isDest
+}: {
+    error: Error;
+    host?: string;
+    log: Logger;
+    isDest?: boolean;
+}): void {
+    log.error(error.message);
+    if (isAxiosError(error) && error.response?.data) {
+        const errorMessage = getErrorMessageFromString(error.response?.data);
+        if (errorMessage) {
+            prettyPrintError({ error: errorMessage, host: host, log: log, isDest: isDest });
+        } else {
+            log.error(error.response.data.toString());
+        }
+    }
+}
+
+/**
+ * Get ErrorMessage object from response contain an error as a string.
+ *
+ * @param data string value
+ * @returns undefined if an error object is not found or populated ErrorMessage object
+ */
+export function getErrorMessageFromString(data: unknown): ErrorMessage | undefined {
+    let error;
+    if (typeof data === 'string') {
+        try {
+            const errorMsg = JSON.parse(data);
+            if (errorMsg.error) {
+                error = errorMsg.error as ErrorMessage;
+            }
+        } catch {
+            // Not much we can do!
+        }
+    }
+    return error;
+}
