+++ packages/axios-extension/src/abap/ui5-abap-repository-service.ts
@@ -1,6 +1,5 @@
 import type { AxiosResponse, AxiosRequestConfig } from 'axios';
-import { prettyPrintError, prettyPrintMessage } from './message';
-import type { ErrorMessage } from './message';
+import { logError, getErrorMessageFromString, prettyPrintError, prettyPrintMessage } from './message';
 import { ODataService } from '../base/odata-service';
 import { isAxiosError } from '../base/odata-request-error';
 /**
@@ -209,7 +208,7 @@ export class Ui5AbapRepositoryService extends ODataService {
                 // Test mode returns a HTTP response code of 403 so we dont want to show all error messages
                 prettyPrintError(
                     {
-                        error: this.getErrorMessageFromString(response?.data),
+                        error: getErrorMessageFromString(response?.data),
                         log: this.log,
                         host: frontendUrl,
                         isDest: this.isDest
@@ -219,7 +218,7 @@ export class Ui5AbapRepositoryService extends ODataService {
             }
             return response;
         } catch (error) {
-            this.logError({ error, host: frontendUrl });
+            logError({ error, host: frontendUrl, log: this.log });
             throw error;
         }
     }
@@ -254,7 +253,7 @@ export class Ui5AbapRepositoryService extends ODataService {
                 return undefined;
             }
         } catch (error) {
-            this.logError({ error, host });
+            logError({ error, host, log: this.log });
             throw error;
         }
     }
@@ -430,45 +429,5 @@ export class Ui5AbapRepositoryService extends ODataService {
                 throw error;
             }
         }
-    }
-
-    /**
-     * Log errors more user friendly if it is a standard Gateway error.
-     *
-     * @param e error thrown by Axios after sending a request
-     * @param e.error error from Axios
-     * @param e.host hostname
-     */
-    protected logError({ error, host }: { error: Error; host?: string }): void {
-        this.log.error(error.message);
-        if (isAxiosError(error) && error.response?.data) {
-            const errorMessage = this.getErrorMessageFromString(error.response?.data);
-            if (errorMessage) {
-                prettyPrintError({ error: errorMessage, host, log: this.log, isDest: this.isDest });
-            } else {
-                this.log.error(error.response.data.toString());
-            }
-        }
-    }
-
-    /**
-     * Get ErrorMessage object from response contain an error as a string.
-     *
-     * @param data string value
-     * @returns undefined if an error object is not found or populated ErrorMessage object
-     */
-    protected getErrorMessageFromString(data: unknown): ErrorMessage | undefined {
-        let error;
-        if (typeof data === 'string') {
-            try {
-                const errorMsg = JSON.parse(data);
-                if (errorMsg.error) {
-                    error = errorMsg.error as ErrorMessage;
-                }
-            } catch {
-                // Not much we can do!
-            }
-        }
-        return error;
     }
 }
